+++ docs/src/release_notes.rst
@@ -7,6 +7,10 @@ This release notes format is based on
 `Keep a Changelog <https://keepachangelog.com>`_.
 Mold adheres to `Semantic Versioning <https://semver.org>`_.
 
+Unreleased
+----------
+- Prefill configuration values
+
 0.1.0 (2021-08-16)
 ------------------
 - Crude initial release